+++ ui/app/actions.js
@@ -117,8 +117,21 @@ var actions = {
   BUY_ETH_SUBVIEW: 'BUY_ETH_SUBVIEW',
   UPDATE_COINBASE_AMOUNT: 'UPDATE_COIBASE_AMOUNT',
   updateCoinBaseAmount: updateCoinBaseAmount,
-  UPDATE_COINBASE_ADDRESS: 'UPDATE_COIBASE_ADDRESS',
-  updateCoinBaseAddress: updateCoinBaseAddress,
+  UPDATE_BUY_ADDRESS: 'UPDATE_BUY_ADDRESS',
+  updateBuyAddress: updateBuyAddress,
+  COINBASE_SUBVIEW: 'COINBASE_SUBVIEW',
+  coinBaseSubview: coinBaseSubview,
+  SHAPESHIFT_SUBVIEW: 'SHAPESHIFT_SUBVIEW',
+  shapeShiftSubview: shapeShiftSubview,
+  PAIR_UPDATE: 'PAIR_UPDATE',
+  pairUpdate: pairUpdate,
+  COIN_SHIFT_REQUEST: 'COIN_SHIFT_REQUEST',
+  coinShiftRquest: coinShiftRquest,
+  SHOW_SUB_LOADING_INDICATION: 'SHOW_SUB_LOADING_INDICATION',
+  showSubLoadingIndication: showSubLoadingIndication,
+  HIDE_SUB_LOADING_INDICATION: 'HIDE_SUB_LOADING_INDICATION',
+  hideSubLoadingIndication: hideSubLoadingIndication,
+
 }
 
 module.exports = actions
@@ -502,6 +515,18 @@ function hideLoadingIndication () {
   }
 }
 
+function showSubLoadingIndication () {
+  return {
+    type: actions.SHOW_SUB_LOADING_INDICATION,
+  }
+}
+
+function hideSubLoadingIndication () {
+  return {
+    type: actions.HIDE_SUB_LOADING_INDICATION,
+  }
+}
+
 function showWarning (text) {
   return this.displayWarning(text)
 }
@@ -614,9 +639,90 @@ function updateCoinBaseAmount (value) {
   }
 }
 
-function updateCoinBaseAddress (value) {
+function updateBuyAddress (value) {
   return {
-    type: actions.UPDATE_COINBASE_ADDRESS,
+    type: actions.UPDATE_BUY_ADDRESS,
     value,
+  }
+}
+
+function coinBaseSubview () {
+  return {
+    type: actions.COINBASE_SUBVIEW,
+  }
+}
+
+function pairUpdate (coin) {
+  return (dispatch) => {
+    dispatch(actions.showSubLoadingIndication())
+    dispatch(actions.hideWarning())
+    shapeShiftRequest('marketinfo', {pair: `${coin.toLowerCase()}_eth`}, (mktResponse) => {
+      dispatch(actions.hideSubLoadingIndication())
+      dispatch({
+        type: actions.PAIR_UPDATE,
+        value: {
+          marketinfo: mktResponse,
+        },
+      })
+    })
+  }
+}
+
+function shapeShiftSubview (network) {
+  var pair
+  network === 'classic' ? pair = 'btc_etc' : pair = 'btc_eth'
+
+  return (dispatch) => {
+    dispatch(actions.showSubLoadingIndication())
+    shapeShiftRequest('marketinfo', {pair}, (mktResponse) => {
+      shapeShiftRequest('getcoins', {}, (response) => {
+        dispatch(actions.hideSubLoadingIndication())
+        if (mktResponse.error) return dispatch(actions.showWarning(mktResponse.error))
+        dispatch({
+          type: actions.SHAPESHIFT_SUBVIEW,
+          value: {
+            marketinfo: mktResponse,
+            coinOptions: response,
+          },
+        })
+      })
+    })
+  }
+}
+
+function coinShiftRquest (data) {
+  return (dispatch) => {
+    dispatch(actions.showSubLoadingIndication())
+    shapeShiftRequest('shift', { method: 'POST', data}, (response) => {
+      dispatch(actions.hideSubLoadingIndication())
+      if (response.error) return dispatch(actions.showWarning(response.error))
+      dispatch({
+        type: actions.COIN_SHIFT_REQUEST,
+        value: {
+          response: response,
+        },
+      })
+    })
+  }
+}
+function shapeShiftRequest (query, options, cb) {
+  var queryResponse, method
+  !options ? options = {} : null
+  options.method ? method = options.method : method = 'GET'
+  var requestListener = function (response){
+    queryResponse = JSON.parse(this.responseText)
+    cb ? cb(queryResponse) : null
+    return queryResponse
+  }
+  var shapShiftReq = new XMLHttpRequest()
+  shapShiftReq.addEventListener('load', requestListener)
+  shapShiftReq.open(method, `https://shapeshift.io/${query}/${options.pair ? options.pair : ''}`, true)
+
+  if (options.method === 'POST') {
+    var jsonObj = JSON.stringify(options.data)
+    shapShiftReq.setRequestHeader('Content-Type', 'application/json')
+    return shapShiftReq.send(jsonObj)
+  } else {
+    return shapShiftReq.send()
   }
 }